# workspace/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: نظام إدارة توزيع السلع - Distribution Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /distributors:
    get:
      operationId: getDistributors
      tags: [distributors]
      summary: Get all distributors
      responses:
        "200":
          description: List of distributors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Distributor"
    post:
      operationId: createDistributor
      tags: [distributors]
      summary: Create distributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDistributorRequest"
      responses:
        "201":
          description: Distributor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
  /distributors/{id}:
    get:
      operationId: getDistributor
      tags: [distributors]
      summary: Get distributor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Distributor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
    put:
      operationId: updateDistributor
      tags: [distributors]
      summary: Update distributor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDistributorRequest"
      responses:
        "200":
          description: Updated distributor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
    delete:
      operationId: deleteDistributor
      tags: [distributors]
      summary: Delete distributor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /distributors/{id}/location:
    put:
      operationId: updateDistributorLocation
      tags: [distributors]
      summary: Update distributor location (GPS tracking)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks
      parameters:
        - name: distributorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /suggestions:
    get:
      operationId: getStoreSuggestions
      tags: [suggestions]
      summary: Get store suggestions from distributors
      responses:
        "200":
          description: Store suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreSuggestion"
    post:
      operationId: createStoreSuggestion
      tags: [suggestions]
      summary: Distributor suggests a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreSuggestionRequest"
      responses:
        "201":
          description: Suggestion submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSuggestion"
  /suggestions/{id}/approve:
    put:
      operationId: approveStoreSuggestion
      tags: [suggestions]
      summary: Admin approve store suggestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestion approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSuggestion"
  /suggestions/{id}/reject:
    put:
      operationId: rejectStoreSuggestion
      tags: [suggestions]
      summary: Admin reject store suggestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestion rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSuggestion"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Distributor:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        username:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        lastSeen:
          type: string
          format: date-time
          nullable: true
        totalTasksCompleted:
          type: integer
        totalAmountCollected:
          type: number
        debt:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - phone
        - isActive
        - username
        - totalTasksCompleted
        - totalAmountCollected
        - debt
        - createdAt
    CreateDistributorRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - firstName
        - lastName
        - phone
        - username
        - password
    UpdateDistributorRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        password:
          type: string
    LocationUpdate:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required:
        - latitude
        - longitude
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        purchasePrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - purchasePrice
        - sellPrice
        - quantity
        - lowStockThreshold
        - isLowStock
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        purchasePrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
      required:
        - name
        - purchasePrice
        - sellPrice
        - quantity
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        purchasePrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
...
    TaskItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - productName
        - quantity
        - price
    Task:
      type: object
      properties:
        id:
          type: integer
        distributorId:
          type: integer
        distributorName:
          type: string
        storeId:
          type: integer
        storeName:
          type: string
        storeLatitude:
          type: number
        storeLongitude:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskItem"
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
        failureReason:
          type: string
          nullable: true
        distance:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - distributorId
        - distributorName
        - storeId
        - storeName
        - storeLatitude
        - storeLongitude
        - items
        - status
        - totalAmount
        - createdAt
    CreateTaskRequest:
      type: object
      properties:
        distributorId:
          type: integer
        storeId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
        notes:
          type: string
      required:
        - distributorId
        - storeId
        - items
    UpdateTaskRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        notes:
          type: string
        failureReason:
          type: string
...
    StoreSuggestion:
      type: object
      properties:
        id:
          type: integer
        distributorId:
          type: integer
        distributorName:
          type: string
        name:
          type: string
        photoUrl:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - distributorId
        - distributorName
        - name
        - latitude
        - longitude
        - status
        - createdAt
    CreateStoreSuggestionRequest:
      type: object
      properties:
        name:
          type: string
        photoUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - name
        - latitude
        - longitude
...