# workspace/lib/api-zod/src/generated/types/createDistributorRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نظام إدارة توزيع السلع - Distribution Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDistributorRequest {
  firstName: string;
  lastName: string;
  phone: string;
  username: string;
  password: string;
}