# workspace/lib/api-zod/src/generated/types/distributor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نظام إدارة توزيع السلع - Distribution Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Distributor {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  isActive: boolean;
  username: string;
  latitude?: number | null;
  longitude?: number | null;
  lastSeen?: Date | null;
  totalTasksCompleted: number;
  totalAmountCollected: number;
  debt: number;
  createdAt: Date;
}