# workspace/lib/api-zod/src/generated/types/storeSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نظام إدارة توزيع السلع - Distribution Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { StoreSuggestionStatus } from "./storeSuggestionStatus";

export interface StoreSuggestion {
  id: number;
  distributorId: number;
  distributorName: string;
  name: string;
  photoUrl?: string | null;
  latitude: number;
  longitude: number;
  status: StoreSuggestionStatus;
  createdAt: Date;
}