# workspace/lib/api-zod/src/generated/types/updateProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نظام إدارة توزيع السلع - Distribution Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductRequest {
  name?: string;
  purchasePrice?: number;
  sellPrice?: number;
  quantity?: number;
  lowStockThreshold?: number;
}