# workspace/lib/api-zod/src/generated/types/updateTaskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * نظام إدارة توزيع السلع - Distribution Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskRequestStatus } from "./updateTaskRequestStatus";

export interface UpdateTaskRequest {
  status?: UpdateTaskRequestStatus;
  notes?: string;
  failureReason?: string;
}